+++ test/test_fetch.js
@@ -10,12 +10,12 @@ let keySpec = [
         name: 'name',
         defaultKey: '',
         extract: request => request.viewName,
-        fallback: notFoundHandler
+        fallback: (variables, request) => Promise.resolve(notFoundHandler(variables, request))
     },
     {
         name: 'method',
         defaultKey: 'GET',
-        fallback: methodNotAllowedHandler
+        fallback: (variables, request) => Promise.resolve(methodNotAllowedHandler(variables, request))
     }
 ];
 
@@ -23,7 +23,7 @@ suite("fakeserver fetch", function() {
     test("responds to GET request", function(done) {
         let server = new FakeFetch(keySpec);
         function handler(variables, request) {
-            return new FakeResponse(JSON.stringify({'animal': variables.id}));
+            return Promise.resolve(new FakeResponse(JSON.stringify({'animal': variables.id})));
         }
         server.register('animals/{id}', handler);
 
@@ -37,7 +37,7 @@ suite("fakeserver fetch", function() {
     test("responds to GET request with body directly", function(done) {
         let server = new FakeFetch(keySpec);
         function handler(variables, request) {
-            return JSON.stringify({'animal': variables.id});
+            return Promise.resolve(JSON.stringify({'animal': variables.id}));
         }
         server.register('animals/{id}', handler);
 
@@ -52,7 +52,7 @@ suite("fakeserver fetch", function() {
     test("404", function(done) {
         let server = new FakeFetch(keySpec);
         function handler(variables, request) {
-            return new FakeResponse(JSON.stringify({'animal': variables.id}));
+            return new Promise.resolve(FakeResponse(JSON.stringify({'animal': variables.id})));
         }
         server.register('animals/{id}', handler);
 
@@ -66,7 +66,7 @@ suite("fakeserver fetch", function() {
     test("POST request", function(done) {
         let server = new FakeFetch(keySpec);
         function handler(variables, request) {
-            return JSON.stringify({'animal': variables.id});
+            return Promise.resolve(JSON.stringify({'animal': variables.id}));
         }
         server.register('animals/{id}', handler, {
             method: 'POST'
@@ -86,7 +86,7 @@ suite("fakeserver fetch", function() {
     test("POST request means no GET", function(done) {
         let server = new FakeFetch(keySpec);
         function handler(variables, request) {
-            return JSON.stringify({'animal': variables.id});
+            return Promise.resolve(JSON.stringify({'animal': variables.id}));
         }
         server.register('animals/{id}', handler, {
             method: 'POST'
@@ -99,6 +99,28 @@ suite("fakeserver fetch", function() {
             done();
         });
     });
+    test("POST request with JSON body", function(done) {
+        let server = new FakeFetch(keySpec);
+        function handler(variables, request) {
+            return request.json().then(JSON.stringify);
+        }
+        server.register('animals/{id}', handler, {
+            method: 'POST'
+        });
 
+        const fetch = server.getFetch();
+
+        fetch(
+            '/animals/chicken',
+            {
+                method: 'POST',
+                body: JSON.stringify({
+                    foo: 'bar'
+                })
+            }).then(json).then(function(data) {
+            assert.deepEqual(data, {'foo': 'bar'});
+            done();
+        });
+    });
 });